+++ frontend/src/routes/index.tsx
@@ -1,11 +1,6 @@
 import { QueryClient } from "@tanstack/react-query";
 import { Explorer } from "ode-explorer/lib";
-import {
-  RouteObject,
-  createBrowserRouter,
-  matchPath,
-  redirect,
-} from "react-router-dom";
+import { RouteObject, createBrowserRouter } from "react-router-dom";
 
 import { explorerConfig } from "~/config/config";
 import Root from "~/routes/root";
@@ -57,23 +52,11 @@ const routes = (queryClient: QueryClient): RouteObject[] => [
   },
   {
     path: "/*",
-    element: <Root />,
-    loader: () => {
-      const ngLocation = location.hash.substring(1);
-
-      const blog = matchPath("/view/:blogId", ngLocation);
-      const post = matchPath("/detail/:blogId/:postId", ngLocation);
-      if (blog) {
-        return redirect(`/id/${blog?.params.blogId}`);
-      }
-      if (post) {
-        return redirect(
-          `/id/${post?.params.blogId}/post/${post?.params.postId}`,
-        );
-      }
-
-      // TODO add 404 page
-      throw new Error("Not found");
+    async lazy() {
+      const { loader } = await import("./old-url");
+      return {
+        loader,
+      };
     },
   },
 ];
